move config common functions to utils
